requirements.txt
openai
llama-index
pypdf
gradio
python-dotenv
flask